# swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "DTApp",
    "version": "0.0.1",
    "description": "Digital Twin Application, unfinished."
  },
  "paths": {
    "/user": {
      "post": {
        "description": "Creates a new user",
        "tags": [
          "user"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created"
          },
          "400": {
            "description": "Other/Unknown"
          },
          "422": {
            "description": "User already exists"
          }
        }
      },
      "patch": {
        "description": "Modify the logged in user",
        "tags": [
          "user"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/User"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Updated"
          },
          "400": {
            "description": "Other/Unknown"
          },
          "422": {
            "description": "New username already exists"
          }
        }
      }
    },
    "/user/token": {
      "post": {
        "description": "Adds a token to user (used for fitbit auth)",
        "tags": [
          "fitbit"
        ],
        "responses": {
          "default": {
            "description": "Object containing the token",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "token": {
                      "type": "string",
                      "description": "The token"
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/withings/auth": {
      "get": {
        "description": "Starts the authentication process for withings api",
        "tags": [
          "withings"
        ],
        "responses": {
          "302": {
            "description": "Redirects to withings website"
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "User": {
        "type": "object",
        "required": [
          "password",
          "username"
        ],
        "properties": {
          "username": {
            "type": "string"
          },
          "password": {
            "type": "string"
          }
        }
      }
    }
  },
  "tags": [
    {
      "name": "user",
      "description": "Everything to do with user"
    },
    {
      "name": "withings",
      "description": "Everything to do with withings"
    }
  ]
}